Se crearon las vistas, modeolos, controladores y rutas  para gregar y ver listado de monedas

+++ database/migrations/2024_09_09_215530_create_monedas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('monedas', function (Blueprint $table) {
             $table->id();
             $table->string('moneda');
-            $table->integer('conversion');
+            $table->decimal('conversion',8,2);
             $table->timestamps();
         });
     }
